+++ src/main/java/com/github/theredbrain/rpginventory/registry/ClientEventsRegistry.java
@@ -44,12 +44,14 @@ public class ClientEventsRegistry {
 				if (stack.isIn(Tags.BELTS)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.belt"));
 				}
-				if (stack.isIn(Tags.HAND_ITEMS) && stack.isIn(Tags.OFFHAND_ITEMS)) {
-					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.both_hands"));
-				} else if (stack.isIn(Tags.HAND_ITEMS)) {
-					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.hand"));
-				} else if (stack.isIn(Tags.OFFHAND_ITEMS)) {
-					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.offhand"));
+				if (!stack.isIn(Tags.TWO_HANDED_ITEMS)) {
+					if (stack.isIn(Tags.HAND_ITEMS) && stack.isIn(Tags.OFFHAND_ITEMS)) {
+						lines.add(Text.translatable("item.additional_tooltip.equipment_slot.both_hands"));
+					} else if (stack.isIn(Tags.HAND_ITEMS)) {
+						lines.add(Text.translatable("item.additional_tooltip.equipment_slot.hand"));
+					} else if (stack.isIn(Tags.OFFHAND_ITEMS)) {
+						lines.add(Text.translatable("item.additional_tooltip.equipment_slot.offhand"));
+					}
 				}
 				if (stack.isIn(Tags.SPELLS)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.spell"));